Fix typed-native proof CI gates
