+++ benchmarks/compiler_output/workloads.toml
@@ -258,6 +258,9 @@ name = "named_region_numeric_loop_has_fp_ops"
 min = { fmul = 1, fadd = 1 }
 detail = "numeric loop FP multiply/add shape"
 
+[workloads.loop_data_dependent.native_rep_checks]
+allow_materialization_reasons = ["runtime_api"]
+
 [workloads.fma_contract]
 source = "benchmarks/compiler_output/fixtures/fma_contract.ts"
 kind = "fma_contract"
@@ -648,6 +651,7 @@ name = "numeric_array_push_dynamic_fallback"
 expr_kind = "NumericArrayPush"
 consumer = "js_array_push_f64"
 access_mode = "dynamic_fallback"
+materialization_reason = "runtime_api"
 
 [[workloads.numeric_arrays.native_rep_checks.require_records]]
 name = "numeric_array_get_fast_f64"
@@ -662,6 +666,7 @@ name = "numeric_array_get_dynamic_fallback"
 expr_kind = "NumericArrayIndexGet"
 consumer = "js_typed_feedback_array_index_get_fallback_boxed"
 access_mode = "dynamic_fallback"
+materialization_reason = "runtime_api"
 
 [[workloads.numeric_arrays.native_rep_checks.require_records]]
 name = "numeric_array_set_fast_f64"
@@ -676,6 +681,7 @@ name = "numeric_array_set_dynamic_fallback"
 expr_kind = "NumericArrayIndexSet"
 consumer_contains = "fallback"
 access_mode = "dynamic_fallback"
+materialization_reason = "runtime_api"
 
 [workloads.raw_numeric_object_fields]
 source = "tests/raw_numeric_object_fields.ts"
